ref(cloudflare): Move internal files and functions around
